requirements.txt
opencv-python
ultralytics
torch
torchvision
pillow
numpy
requests
customtkinter